requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
exceptiongroup==1.2.2
flake8==7.1.1
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.3
pytest-django==4.9.0
python-decouple==3.8
sqlparse==0.5.2
tomli==2.1.0
typing_extensions==4.12.2
